updated README.md and Github Actions tests
